requirements.txt
Flask
#Flask==2.1.1
# Flask-SocketIO==5.1.1  # If using WebSocket for real-time communication
# Flask-SSE==0.2.0  # If using Server-Sent Events for real-time communication
mysql-connector